requirements.txt
    accelerate==1.8.1
    annotated-types==0.7.0
    anyio==4.9.0
    audioread==3.0.1
    black==24.3.0
    certifi==2025.7.9
    cffi==1.17.1
    cfgv==3.4.0
    charset-normalizer==3.4.2
    click==8.1.7
    decorator==5.2.1
    distlib==0.3.9
    docstring_parser_fork==0.0.12
    fastapi==0.116.0
    ffmpeg==1.4
    filelock==3.13.1
    flake8==7.0.0
    fsspec==2024.6.1
    h11==0.16.0
    hf-xet==1.1.5
    huggingface-hub==0.33.2
    HyperPyYAML==1.2.2
    identify==2.6.12
    idna==3.10
    iniconfig==2.1.0
    isort==5.13.2
    Jinja2==3.1.4
    joblib==1.5.1
    lazy_loader==0.4
    librosa==0.11.0
    llvmlite==0.44.0
    MarkupSafe==2.1.5
    mccabe==0.7.0
    more-itertools==10.7.0
    mpmath==1.3.0
    msgpack==1.1.1
    mypy_extensions==1.1.0
    networkx==3.3
    nodeenv==1.9.1
    numba==0.61.2
    numpy==2.1.2
    openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
    opencv-python==4.12.0.88
    packaging==25.0
    pandas==2.3.1
    pathspec==0.12.1
    pillow==11.0.0
    platformdirs==4.3.8
    pluggy==1.6.0
    pooch==1.8.2
    pre_commit==4.2.0
    psutil==7.0.0
    pycodestyle==2.11.0
    pycparser==2.22
    pydantic==2.11.7
    pydantic_core==2.33.2
    pydoclint==0.4.1
    pyflakes==3.2.0
    pytest==7.4.0
    python-dateutil==2.9.0.post0
    python-multipart==0.0.20
    pytz==2025.2
    PyYAML==6.0.2
    regex==2024.11.6
    requests==2.32.4
    ruamel.yaml==0.18.14
    ruamel.yaml.clib==0.2.12
    safetensors==0.5.3
    scikit-learn==1.7.0
    scipy==1.15.0
    sentencepiece==0.2.0
    setuptools==70.2.0
    six==1.17.0
    sniffio==1.3.1
    soundfile==0.13.1
    soxr==0.5.0.post1
    speechbrain==1.0.3
    starlette==0.46.2
    sympy==1.13.3
    threadpoolctl==3.6.0
    tiktoken==0.9.0
    tokenizers==0.21.2
    torch==2.7.1
    torchaudio==2.7.1
    torchvision==0.22.1
    tqdm==4.67.1
    transformers==4.53.1
    typing-inspection==0.4.1
    typing_extensions==4.12.2
    tzdata==2025.2
    urllib3==2.5.0
    uvicorn==0.35.0
    virtualenv==20.31.2
    yamllint==1.35.1
    fastapi
    fastapi
    uvicorn[standard]
    transformers
    torch
    pillow
    requests
    opencv-python
    pydantic
    # Optional runtime helpers
    accelerate
    safetensors

